Fix line select and stop times bugs.
